requirements.txt
django~=4.2.3
psycopg2-binary
djangorestframework~=3.14.0
django-filter~=23.2
djangorestframework-simplejwt
django-cors-headers
google-cloud-storage
gunicorn
requests
